Support for http proxy
When `proxy_info` is supplied, a connection is first established
with the proxy server and a `CONNECT` request is issued. Assuming
the proxy responds with a 200, all the headers are consumed so that
the socket is ready to be used as if it were a direct connection
to the websocket server.

Supports the Basic Authentication Scheme (RFC-7617). If a
user/password combination is given in the proxy url, a suitable
`Proxy-Authentication` will be sent.